flask/requirements.txt
flask
flask-cors
python-dotenv
Flask-SQLAlchemy